services/api-gateway/requirements.txt
# =============================================================================
# CONECTA PLUS - API Gateway Dependencies
# =============================================================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-multipart==0.0.6
starlette==0.35.1

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Database
asyncpg==0.29.0
redis==5.0.1

# QR Code e PDF
qrcode[pil]==7.4.2
reportlab==4.0.8
Pillow==10.2.0
python-barcode==0.15.1
openpyxl==3.1.2

# Email
aiosmtplib==3.0.1

# WebSocket
websockets==12.0

# ML/NLP (optional, graceful fallback)
numpy>=1.24.0
# scikit-learn>=1.3.0  # Uncomment if ML features needed

# Testes
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilitários
python-dateutil==2.8.2
jinja2==3.1.3
python-dotenv==1.0.0

# Production
gunicorn==21.2.0
